Moved get_chapter_nextprev function to helpers
